change symboltable address from int to short

+++ assembler/SymbolTable.cpp
@@ -12,13 +12,13 @@ namespace hack_assembler {
 SymbolTable::SymbolTable() : symbol_table_() {}
 
 SymbolTable::SymbolTable(
-    std::initializer_list<std::pair<Symbol, int>> symbol_pairs) {
+    std::initializer_list<std::pair<Symbol, short>> symbol_pairs) {
   for (const auto& [symbol, address] : symbol_pairs) {
     symbol_table_.emplace(symbol, address);
   }
 }
 
-void SymbolTable::addEntry(const Symbol& symbol, int address) {
+void SymbolTable::addEntry(const Symbol& symbol, short address) {
   symbol_table_.emplace(symbol, address);
 }
 
@@ -26,7 +26,7 @@ bool SymbolTable::contains(const Symbol& symbol) const {
   return symbol_table_.contains(symbol);
 }
 
-int SymbolTable::getAddress(const Symbol& symbol) const {
+short SymbolTable::getAddress(const Symbol& symbol) const {
   return symbol_table_.at(symbol);
 }
